Rename blockchainshim interface
