+++ protocol/interface.go
@@ -1,9 +1,30 @@
 package protocol
 
 import (
+	"math/big"
+
 	"github.com/dogechain-lab/dogechain/blockchain"
 	"github.com/dogechain-lab/dogechain/helper/progress"
+	"github.com/dogechain-lab/dogechain/types"
 )
+
+// Blockchain is the interface required by the syncer to connect to the blockchain
+type blockchainShim interface {
+	SubscribeEvents() blockchain.Subscription
+	Header() *types.Header
+	CurrentTD() *big.Int
+
+	GetTD(hash types.Hash) (*big.Int, bool)
+	GetReceiptsByHash(types.Hash) ([]*types.Receipt, error)
+	GetBodyByHash(types.Hash) (*types.Body, bool)
+	GetHeaderByHash(types.Hash) (*types.Header, bool)
+	GetHeaderByNumber(n uint64) (*types.Header, bool)
+
+	// advance chain methods
+	WriteBlock(block *types.Block) error
+	VerifyFinalizedBlock(block *types.Block) error
+	CalculateGasLimit(number uint64) (uint64, error)
+}
 
 type Progression interface {
 	// StartProgression starts progression
